Abstract
Starting with the recognition of dissatisfaction with Cheonan buses, we focused on the civil service factor called 'Failure of Dispatch Gap' among the civil service factors. Not only the feeling but also the analysis of whether Cheonan buses are actually implementing the designated interval. Hypothesis and analysis that 'time zones' will affect the fulfillment of the dispatch interval..